Added Material16 and Material32 types. Small fixes for where 8-bit material was assumed.

+++ library/PolyVoxCore/include/PolyVoxCore/CubicSurfaceExtractor.h
@@ -63,7 +63,7 @@ namespace PolyVox
 		void execute();		
 
 	private:
-		int32_t addVertex(float fX, float fY, float fZ, uint8_t uMaterial, Array<3, IndexAndMaterial>& existingVertices);
+		int32_t addVertex(float fX, float fY, float fZ, uint32_t uMaterial, Array<3, IndexAndMaterial>& existingVertices);
 		bool performQuadMerging(std::list<Quad>& quads);
 		bool mergeQuads(Quad& q1, Quad& q2);
